% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_methods.R
\name{cluster_discretemixture}
\alias{cluster_discretemixture}
\title{Mixed discrete and continuous latent trait}
\usage{
cluster_discretemixture(data, k = 1:10, verbose = FALSE, bic_plot = FALSE)
}
\description{
Mixed discrete and continuous latent trait
}
\examples{
data(veronica)
v2 <- data.frame(lapply(veronica, as.factor))
set.seed(887766)
cl <- cluster_discretemixture(v2, verbose=TRUE, bic_plot=TRUE)
}
