% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_methods.R
\name{cluster_gaussianmixture}
\alias{cluster_gaussianmixture}
\title{Gaussian mixture}
\usage{
cluster_gaussianmixture(data, k = 1:10, verbose = TRUE, bic_plot = FALSE)
}
\description{
Gaussian mixture
}
\examples{
data(oliveoil)
olive <- oliveoil[, 3:10]
gmix <- cluster_gaussianmixture(olive, k=1:10, verbose=TRUE)
# here higher bic is better
table(gmix$classification)
}
