% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_methods.R
\name{cluster_hier}
\alias{cluster_hier}
\title{Hierarchical clustering}
\usage{
cluster_hier(d, k = NULL, method = c("average", "single", "complete"))
}
\description{
Average linkage by default other wise set method
If k is specified cutree is
}
\examples{

# jaccard on veronica, average link
d <- dist_jaccard(veronica)
plot(cluster_hier(d)) # average linkage by default
cluster_hier(d, k=8)

# euclidean on scaled geyser, average link
d <- dist_eucl(scale(geyser))
hcl <- cluster_hier(d, k=4)
par(mfrow=c(1,2))
plot(hcl$hclus)
plot_cluster(scale(geyser), hcl$cluster)

}
