% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_methods.R
\name{cluster_kmeans}
\alias{cluster_kmeans}
\title{K-means with better defaults}
\usage{
cluster_kmeans(x, k, nstart = 100, iter.max = 100)
}
\description{
K-means with better defaults
}
\examples{

set.seed(1)
km <- cluster_kmeans(clusterdata1, k=3)
plot_cluster(clusterdata1, km$cluster)
cluster_profiles(clusterdata1, km$cluster)

}
