% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_methods.R
\name{cluster_pam}
\alias{cluster_pam}
\title{PAM}
\usage{
cluster_pam(d, k)
}
\arguments{
\item{d}{data or distance (if data euclidean distance is used)}

\item{k}{number of clusters}
}
\description{
K-means on distance to medoids (object)
}
\examples{
data(bundestag)
b <- bundestag[1:5]
d <- dist_manhattan(b)
set.seed(1)
p <- lbclus::pam(d=d, k=5)

}
