% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{cluster_profiles}
\alias{cluster_profiles}
\title{compute cluster profiles (means)}
\usage{
cluster_profiles(data, clvec)
}
\description{
compute cluster profiles (means)
}
\examples{
cluster_profiles(airquality[, -(5:6)], airquality$Month)
}
