% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_methods.R
\name{cluster_skewtmixture}
\alias{cluster_skewtmixture}
\title{Skew-T mixture}
\usage{
cluster_skewtmixture(data, k = 1:10, verbose = TRUE)
}
\arguments{
\item{data}{data}

\item{k}{number of cluster to search for, optimized for BIC (lower better)}

\item{verbose}{show best result parameter estimates}
}
\description{
Skew-T mixture
}
\examples{
data(oliveoil)
olive <- oliveoil[, 3:10]
solive <- scale(olive)
stolive <- cluster_skewtmixture(solive)
# here smaller BIC is better

}
