% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_gower}
\alias{dist_gower}
\title{Gower distance}
\usage{
dist_gower(x, type = list())
}
\description{
Gower distance
}
\examples{

data(housing)
head(housing) # chas (4-th) is binary, rad (9-th) is ordinal
summary(housing)

# binary: simple matching distance (symm)
# binary: jaccard (asymm)
# nominal: simple matching distance (symm). Factor are handled like that
# ordered: coded with integer and progressive and treated as numerical
# unspecified: numerical with L_1

# utility function
sneakpeek <- function(x) { sel <- 1:5 ; as.matrix(x)[sel, sel] }

# all numerical (L1)
d0 <- dist_gower(housing)
sneakpeek(d0) # it complains

# chas with simple matching distance
typelist1 <- list(symm=4)
d1 <- dist_gower(housing, typelist1)
sneakpeek(d1)

# chas and rad with simple matching distance (4th) rad (9th)
typelist2 <- list(symm=c(4,9))
d2 <- dist_gower(housing, typelist2) # this doesnt' work
housing$rad <- as.factor(housing$rad) # factor should be simple matching distanced 
d2b <- dist_gower(housing, typelist1)
sneakpeek(d2b)
d2c <- dist_gower(housing, typelist2) # doesn't work...

# chas (4th) with simple matching distance and ordering rad (9th)
housing$rad <- as.ordered(housing$rad)
d3 <- dist_gower(housing, typelist1)
sneakpeek(d3) # no changes with d1 bc ordered are basically L1 treated

# Take home messages: just ..
# - change to factor or ordered as appropriate
# - for binary variable choose jaccard or simple matching distance
}
