% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_mahalanobis}
\alias{dist_mahalanobis}
\title{Squared mahalanobis distance}
\usage{
dist_mahalanobis(x)
}
\description{
Squared mahalanobis distance
}
\examples{
data(oliveoil)
olive <- oliveoil[, 3:10]
dm <- dist_mahalanobis(olive)
sel <- 1:5
as.matrix(dm)[sel, sel]

}
