% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_smd}
\alias{dist_smd}
\title{Simple matching distance}
\usage{
dist_smd(x)
}
\description{
Simple matching distance
}
\examples{
a <- dist_smd(veronica)
b <- dist(veronica, method = "manhattan")/ncol(veronica) # just in binary case it's the same
sel <- 1:5
as.matrix(a)[sel, sel]
as.matrix(b)[sel, sel]
}
