% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_k.R
\name{gapnc}
\alias{gapnc}
\title{gapnc function with minimal changes for personal use/automation}
\usage{
gapnc(
  data,
  FUNcluster = lbclus::cluster_kmeans,
  K.max = 10,
  B = 100,
  d.power = 2,
  spaceH0 = "scaledPCA",
  method = "globalSEmax",
  SE.factor = 2,
  view = TRUE,
  ...
)
}
\description{
use by default kmeans with better defaults (eg iter.max
}
\examples{
data(clusterdata2)
set.seed(112)
gap_clus2 <- gapnc(clusterdata2, K.max = 15) # optimal: 9 or 10
}
