% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{index_ari}
\alias{index_ari}
\title{Adjusted Rand Index (agreement of clustering solutions like CCC)}
\usage{
index_ari(cl_x, cl_y)
}
\description{
Adjusted Rand Index (agreement of clustering solutions like CCC)
}
\examples{
olive <- oliveoil[-c(1:2)]
solive <- scale(olive)
olive3 <- lbclus::kmeans(olive, 3)
olive3s <- lbclus::kmeans(solive, 3)

# agreement between clustering solutions
index_ari(olive3$cluster, olive3s$cluster)

# comparison with known classification (higher better)
index_ari(olive3$cluster, oliveoil$macro.area)
index_ari(olive3s$cluster, oliveoil$macro.area)

}
