% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{index_asw}
\alias{index_asw}
\title{Average Silhouette Width}
\usage{
index_asw(cl, d)
}
\arguments{
\item{cl}{a classification vector (integers) or a list with $clustering component}

\item{d}{dist object}
}
\description{
Quality of clustering solutions given distance between units (the higher the
better)
}
\examples{

data(bundestag)
b <- bundestag[1:5]
d <- dist_manhattan(b)
bpam <- lbclus::pam(d, k=2)

index_asw(bpam$cluster, d=d) # should be 0.48
index_asw(bpam, d=d) #same?

# choosing n of clusters within k-means
# -------------------------------------
data(clusterdata1)
plot(clusterdata1)
sdat <- scale(clusterdata1)
sdat_dist <- dist(sdat)

K <- 2:5
names(K) <- K
clusters <- lapply(K, function(k) lbclus::kmeans(sdat, k=k)$cluster)
sapply(clusters, function(cl) index_asw(cl = cl, d = sdat_dist))

}
