% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{index_asw_compare}
\alias{index_asw_compare}
\title{Average Silhouette compare}
\usage{
index_asw_compare(lcl, d)
}
\arguments{
\item{lcl}{list of clustering solutions}

\item{d}{dist object of data}
}
\description{
Compare different clustering solutions in term of quality, given a common
distances description of data
}
\examples{

# exercise 5
data(glass)
sglass <- scale(glass)
d <- dist_eucl(sglass)

km3 <- cluster_kmeans(sglass, 2)
gm <- cluster_gaussianmixture(sglass, k = 4:6, verbose=FALSE)
l <- list("km3"= km3$cluster, "gm"=gm$classification)

index_asw_compare(lcl=l,d=d)
}
