% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{infoloss_mds}
\alias{infoloss_mds}
\title{Quantifying information loss (stress) using MDS derived variables}
\usage{
infoloss_mds(d, ndim = 1:10)
}
\description{
Useful for obtaining data (choosing a small p) for following clustering not having a
too much high info loss
}
\examples{
data(veronica)
d <- dist_jaccard(veronica)
infoloss_mds(d)
}
