% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{infoloss_pca}
\alias{infoloss_pca}
\title{Quantifying information loss (stress) using PCA derived variables}
\usage{
infoloss_pca(x, scale = TRUE, verbose = FALSE)
}
\description{
Useful for obtaining data (choosing a small p) for following clustering not
having a too much high info loss
}
\examples{
data(oliveoil)
olive <- oliveoil[, 3:10]
infoloss_pca(olive, scale = FALSE)
infoloss_pca(olive, scale = TRUE)
}
