% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_methods.R
\name{pam}
\alias{pam}
\title{PAM}
\usage{
pam(d, k)
}
\arguments{
\item{d}{data or distance (if data euclidean distance is used)}

\item{k}{number of clusters}
}
\description{
K-means on distance to medoids (object)
}
\examples{

b <- bundestag[1:5]
d <- dist_manhattan(bundestag)
set.seed(1)
lbclus::pam(d=d, k=5)

}
