% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Function for plotting clusters}
\usage{
plot_cluster(coords, clvec, col_alpha = 1, ...)
}
\description{
Function for plotting clusters
}
\examples{
example(cluster_kmeans)
}
