% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{plot_pairs}
\alias{plot_pairs}
\title{pairs with smoothing, histograms and less/more visible correlations}
\usage{
plot_pairs(data, pch = 20, col = lbmisc::col2hex("black", 0.3), ...)
}
\arguments{
\item{data}{dataframe}

\item{pch}{point type}

\item{col}{color}

\item{...}{other parameters given to pairs}
}
\description{
pairs with smoothing, histograms and less/more visible correlations
}
\examples{
pairsplot(airquality, col=airquality[, 5])
}
