% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{cluster_profiles}
\alias{cluster_profiles}
\title{compute cluster profiles (means)}
\usage{
cluster_profiles(data, clvec)
}
\description{
compute cluster profiles (means)
}
\examples{
# example1
example(cluster_kmeans)
# example2 (bad but quick example)
airquality2 <- airquality
airquality2$Month <- factor(airquality2$Month)
airquality2$Day <- factor(airquality2$Day)
cluster_profiles(airquality2, airquality2$Month) # suppose month is result of a clustering
}
