% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{coords_mds}
\alias{coords_mds}
\title{MDS 2-d coordinates}
\usage{
coords_mds(d, verbose = TRUE)
}
\description{
MDS 2-d coordinates
}
\examples{
data(bundestag)
# bundestag example
b <- bundestag[1:5]
d <- dist_corr1(b)
coords_mds(d)

# veronica example
data(veronica)
d <- dist_jaccard(veronica)
plot(coords_mds(d), asp=1)
}
