% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapnc.R
\name{gapnc}
\alias{gapnc}
\title{gapnc function with minimal changes for personal use/automation}
\usage{
gapnc(
  data,
  FUNcluster = lbclus::kmeans,
  K.max = 10,
  B = 100,
  d.power = 2,
  spaceH0 = "scaledPCA",
  method = "globalSEmax",
  SE.factor = 2,
  ...
)
}
\description{
use by default kmeans with better defaults (eg iter.max
}
\examples{

set.seed(112)
gap_clus2 <- gapnc(clusterdata2, K.max = 15)
plot(gap_clus2)
# optimal number of cluster is 9 o 10 a seconda di come si sveglia

}
