% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{index_mean_purity}
\alias{index_mean_purity}
\title{Mean purity index}
\usage{
index_mean_purity(ref, cl)
}
\description{
Used in homework 1 to compare clustering solutions to a known classification
in a supervised approach: the index close to 1 suggest that clustering is
reproducing the reference/known classification better
}
\examples{
library(lbdatasets)
# oliveoil <- read.table("data/oliveoil.dat", header=TRUE)
head(oliveoil)
olive <- oliveoil[-c(1:2)]
cl <- kmeans(olive, 9)$cluster
index_mean_purity(oliveoil$region, cl)

}
